Enable dev "make run" twice, copy installer first
Running "make run" in the base project for a second time used to fail
because re-appending the data files to the already-modified binary is
not possible.

Use the approach from the builder Makefile for the dev Makefile as
well and duplicate the naked installer before appending. This way the
output installer is overwritten with the naked installer every time.