Fix uninstallation of broken symlinks
-f will check if the _target_ of a symlink is a file. This check
coincidentally works for most cases when the target still exists and
is a file. But if the link is broken or the target is anything else
but a file, -f will return false for a symlink and thus not remove it.

Check for type symlink explicitly and remove this uncertainty.