Add go.rice/rice cmd to vendor, again
When vendoring tooling binaries, which are not go modules, import them
temporarily in a separate and otherwise disabled .go file.
`go mod ...` does include these imports in the dependencies, while
`go build ...` errors because they aren't actually importable.
In this case it's the sources for the "rice" command from go.rice.

Create a Makefile recipe, "vendor-refresh", that temporarily imports
this tool, and then recreates and updates the vendored dependencies.
Afterwards disable the import again so the regular build succeeds.

Now the vendor directory can be more easily updated in the future,
without dropping the rice command dependency again.